+++ packages/nft-renderer/panda.config.ts
@@ -15,8 +15,9 @@ export default defineConfig({
   theme: {
     extend: {},
   },
+  outExtension: 'js',
 
   // The output directory for your css system
-  outdir: 'styled-system',
+  outdir: 'src/styled-system',
   jsxFramework: 'react',
 });