+++ packages/rmrk-hooks/src/lib/hooks/use-fetch-ipfs-metadatas.ts
@@ -12,7 +12,7 @@ type Props = {
   ipfsGatewayUrl?: string;
 };
 
-type Options = { enabled?: boolean };
+type Options = { enabled?: boolean; shouldSanitizeIpfsUrls?: boolean };
 
 /**
  * A custom hook that fetches IPFS metadata for the given URIs.
@@ -32,18 +32,19 @@ export const useFetchIpfsMetadatas = (
   options?: Options,
 ) => {
   const rmrkConfig = useRMRKConfig();
-  const { enabled = true } = options || {};
+  const { enabled = true, shouldSanitizeIpfsUrls = true } = options || {};
 
   const results = useQueries({
     queries: (metadataUris || []).map((metadataUri) => ({
       queryKey: ['fetchIpfsMetadata', metadataUri],
       queryFn: () =>
-        fetchIpfsMetadata(
-          metadataUri,
-          ipfsGatewayUrl ||
+        fetchIpfsMetadata(metadataUri, {
+          ipfsGatewayUrl:
+            ipfsGatewayUrl ||
             rmrkConfig?.ipfsGateway ||
             DEFAULT_IPFS_GATEWAY_URLS[DEFAULT_IPFS_GATEWAY_KEYS.cloudflare],
-        ),
+          shouldSanitizeIpfsUrls,
+        }),
       enabled,
     })),
   });