+++ packages/types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rmrk-team/types
 
+## 0.1.1
+
+### Patch Changes
+
+- [`a1cfe3b`](https://github.com/rmrk-team/rmrk-js/commit/a1cfe3b5d7b19ba5eac03c1be3a475753053b425) Thanks [@Yuripetusko](https://github.com/Yuripetusko)! - Improve esm export and update all deps
+
 ## 0.1.0
 
 ### Minor Changes